split test_irf_utils into weight/unweighted
